Move Google Gen AI fixture to allow reuse
